requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database & Cache
supabase==2.10.0
redis==5.2.0

# Security & Cryptography
cryptography==44.0.0
bcrypt==4.2.1
PyJWT==2.10.0
python-jose[cryptography]==3.3.0

# Async Task Queue
arq==0.26.0

# Email
aiosmtplib==3.0.2
email-validator==2.2.0

# Telegram Bot
python-telegram-bot==21.7

# Templating
jinja2==3.1.4

# HTTP Client
httpx==0.28.0
aiohttp==3.11.0

# Security Headers
secure==1.0.1

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0

# Code Quality
ruff==0.8.0
mypy==1.13.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.18